chore(wizard, accordion, stack-view): move common file HeadingLevel to utils
