+++ projects/angular/src/accordion/accordion-panel.spec.ts
@@ -11,12 +11,12 @@ import { ReactiveFormsModule } from '@angular/forms';
 import { By } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 
-import { IfExpandService } from '../utils/conditional/if-expanded.service';
-import { HeadingLevel } from '../wizard';
 import { ClrAccordionPanel } from './accordion-panel';
 import { ClrAccordionModule } from './accordion.module';
 import { AccordionPanelModel } from './models/accordion.model';
 import { AccordionService } from './providers/accordion.service';
+import { IfExpandService } from '../utils/conditional/if-expanded.service';
+import { HeadingLevel } from '../utils/types/heading-level';
 
 @Component({
   template: `
