+++ projects/angular/src/accordion/accordion-panel.ts
@@ -24,14 +24,14 @@ import {
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 
-import { IfExpandService } from '../utils/conditional/if-expanded.service';
-import { ClrCommonStringsService } from '../utils/i18n/common-strings.service';
-import { uniqueIdFactory } from '../utils/id-generator/id-generator.service';
-import { HeadingLevel } from '../wizard';
 import { ClrAccordionDescription } from './accordion-description';
 import { AccordionPanelModel } from './models/accordion.model';
 import { AccordionService } from './providers/accordion.service';
 import { panelAnimation } from './utils/animation';
+import { IfExpandService } from '../utils/conditional/if-expanded.service';
+import { ClrCommonStringsService } from '../utils/i18n/common-strings.service';
+import { uniqueIdFactory } from '../utils/id-generator/id-generator.service';
+import { HeadingLevel } from '../utils/types/heading-level';
 
 @Component({
   selector: 'clr-accordion-panel',
