+++ projects/angular/src/data/stack-view/stack-block.ts
@@ -20,7 +20,7 @@ import {
 import { ClrStackViewLabel } from './stack-view-custom-tags';
 import { ClrCommonStringsService } from '../../utils/i18n/common-strings.service';
 import { uniqueIdFactory } from '../../utils/id-generator/id-generator.service';
-import { HeadingLevel } from '../../wizard';
+import { HeadingLevel } from '../../utils/types/heading-level';
 
 @Component({
   selector: 'clr-stack-block',
