+++ projects/angular/src/wizard/wizard-page-title.ts
@@ -7,7 +7,7 @@
 
 import { Directive, Input, TemplateRef } from '@angular/core';
 
-import { HeadingLevel } from './heading-level';
+import { HeadingLevel } from '../utils/types/heading-level';
 
 @Directive({
   selector: '[clrPageTitle]',
